refactor: Base 64 validator functions
