+++ src/core/utils/index.ts
@@ -1,0 +1,5 @@
+export {
+    isBase64,
+    isValidBase64,
+    validateBase64DataURL
+} from "./base64"
