+++ src/modules/Article/views/CreateOrEdit/composables/useImage.ts
@@ -1,6 +1,7 @@
 import { ref } from 'vue'
 import { Http as h } from "@/core/utils/Http";
 import { productInit } from "@/core/utils/Http/init";
+import { validateBase64DataURL } from "@/core/utils"
 import type { Ref } from "vue";
 import type { ArticleDetail } from "@/modules/Article/types/Article/ArticleDetail";
 import type { ImageType,  Base64} from "@/modules/Article/types/Image";
@@ -34,47 +35,7 @@ export default () => {
     const promise = await convertImageToBase64(url.replace('storage/', ''));
     
     base64Images.value.unshift(promise as unknown as Base64<ImageType>)
-  }
-
-  /**
-   * Important: It's crucial to note that these functions can't guarantee 100% that a string is Base64, but they do provide robust verification. In some cases, malformed strings may pass character validation but fail decoding.
-   * Alternative (Node.js): If you're working with Node.js, you can use the Buffer module to verify the string.
-   * 
-   * // Node.js
-   * function isBase64Node(str) {
-   *   try {
-   *     const buffer = Buffer.from(str, 'base64');
-   *     return str === buffer.toString('base64');
-   *   } catch (error) {
-   *     return false;
-   *   }
-   * }
-   */
-
-  function isBase64(str: string) {
-     const regex = /^[A-Za-z0-9+/]*={0,2}$/;
-     return regex.test(str);
-  }
-
-  function isValidBase64(str: string) {
-     try {
-       const decoded = atob(str);
-       const encoded = btoa(decoded);
-       return str === encoded;
-     } catch (error) {
-       return false;
-     }
-  }
-
-  function validateBase64DataURL(dataURL: string) {
-    const base64Regex = /^data:image\/[a-zA-Z]+;base64,(.+)$/;
-    const match = dataURL.match(base64Regex);
-    if (match) {
-        const base64String = match[1];
-        return isBase64(base64String) && isValidBase64(base64String);
-    }
-    return false;
-  }
+  }  
 
   const addImagePath = (articleDetail: string) => {
     if (imagePaths.value.includes(articleDetail)) {
@@ -116,7 +77,4 @@ export default () => {
     removeImagePath,
     removeAllImagePaths
   }
-
 }
-
-
